Inject 'use strict' directives
ESM always runs in strict mode, but the context where the code is
evaluated might not. Now a `'use strict'` directive is injected at the
top of the file.

If a `'use strict'` directive is already present, it’s reused. Any
nested `'use strict'` directives are redundant and are removed.